Rename NumberGame.py to app/main.py
